osdp/requirements.txt
aws-cdk-lib==2.179.0
constructs>=10.0.0,<11.0.0
aws-cdk.aws-amplify-alpha==2.178.2a0
boto3>=1.28.0
